[Client/*] Rename mgmt -> forms, update packaging
