+++ client/angular.json
@@ -114,17 +114,17 @@
         }
       }
     },
-    "@portal/mgmt": {
-      "root": "libs/mgmt",
-      "sourceRoot": "libs/mgmt",
+    "@portal/forms": {
+      "root": "libs/forms",
+      "sourceRoot": "libs/forms",
       "projectType": "library",
       "prefix": "",
       "architect": {
         "build": {
           "builder": "@angular-devkit/build-ng-packagr:build",
           "options": {
-            "tsConfig": "libs/mgmt/tsconfig.json",
-            "project": "libs/mgmt/ng-package.json"
+            "tsConfig": "libs/forms/tsconfig.json",
+            "project": "libs/forms/ng-package.json"
           }
         },
         "lint": {
@@ -132,7 +132,7 @@
           "options": {
             "tsConfig": [
               "tsconfig.json",
-              "libs/mgmt/tsconfig.json"
+              "libs/forms/tsconfig.json"
             ],
             "exclude": [
               "**/node_modules/**"
