+++ client/libs/forms/ng-package.json
@@ -1,0 +1,25 @@
+{
+  "$schema": "../../node_modules/ng-packagr/ng-package.schema.json",
+  "dest": "../../target/@portal/forms",
+  "lib": {
+    "entryFile": "index.ts",
+    "umdModuleIds": {
+      "@angular-devkit/build-ng-packagr": "0.11.2",
+      "@angular/animations": "7.1.2",
+      "@angular/cdk": "7.1.1",
+      "@angular/common": "7.1.2",
+      "@angular/cli": "7.1.2",
+      "@angular/compiler": "7.1.2",
+      "@angular/compiler-cli": "7.1.2",
+      "@angular/core": "7.1.2",
+      "@angular/forms": "7.1.2",
+      "@angular/material": "7.1.1",
+      "@angular/platform-browser": "7.1.2",
+      "@portal/core": "0.0.3",
+      "ng-packagr": "4.4.5",
+      "rxjs": "6.3.3",
+      "tsickle": "0.34.0",
+      "typescript": "3.1.6"
+    }
+  }
+}
