--- client/libs/mgmt/ng-package.json
@@ -1,17 +1,0 @@
-{
-  "$schema": "../../node_modules/ng-packagr/ng-package.schema.json",
-  "dest": "../../target/@portal/mgmt",
-  "lib": {
-    "entryFile": "index.ts",
-    "umdModuleIds": {
-      "@angular/common": "7.1.1",
-      "@angular/core": "7.1.1",
-      "@angular-devkit/build-ng-packagr": "0.11.0",
-      "@angular/material": "7.1.0",
-      "ng-packagr": "4.4.0",
-      "rxjs": "6.3.3",
-      "tsickle": "0.34.0",
-      "tslib": "1.9.3"
-    }
-  }
-}
