--- client/libs/mgmt/package.json
@@ -1,25 +1,0 @@
-{
-  "name": "@portal/mgmt",
-  "version": "0.0.3",
-  "description": "Portal Client Management Library",
-  "homepage": "https://github.com/codeschluss/wupportal",
-  "author": {
-    "name": "Wupportal Team"
-  },
-  "bugs": {
-    "email": "info@codeschluss.de",
-    "url": "https://github.com/codeschluss/wupportal/issues"
-  },
-  "dependencies": {
-    "@angular/common": "7.1.1",
-    "@angular/core": "7.1.1",
-    "@angular/material": "7.1.0",
-    "rxjs": "6.3.3"
-  },
-  "devDependencies": {
-    "@angular-devkit/build-ng-packagr": "0.11.0",
-    "ng-packagr": "4.4.0",
-    "tsickle": "0.34.0",
-    "tslib": "1.9.3"
-  }
-}
