+++ client/package.json
@@ -55,11 +55,7 @@
     "@ngx-pwa/local-storage": "7.2.1",
     "@ngx-translate/i18n-polyfill": "1.0.0",
     "@portal/core": "file:libs/core",
-<<<<<<< HEAD
-    "@portal/forms": "file:libs/forms",
-=======
     "@portal/mgmt": "file:libs/mgmt",
->>>>>>> Draft mgmt module
     "classlist.js": "1.1.20150312",
     "core-js": "2.6.0",
     "hammerjs": "2.0.8",
